+++ CHANGELOG.md
@@ -15,7 +15,7 @@ FEATURES:
 
 IMPROVEMENTS:
 
-  - Error out if `redir` is not installed but port forwarding was configured [#112](https://github.com/fgrehm/vagrant-lxc/issues/112)
+  - Error out if dependencies are not installed [#11](https://github.com/fgrehm/vagrant-lxc/issues/11) / [#112](https://github.com/fgrehm/vagrant-lxc/issues/112)
   - Support for specifying host interface/ip for binding `redir` [#76](https://github.com/fgrehm/vagrant-lxc/issues/76)
   - Add Vagrantfile VM name to the container name [#115](https://github.com/fgrehm/vagrant-lxc/issues/115)
   - Properly handle forwarded port collisions [#5](https://github.com/fgrehm/vagrant-lxc/issues/5)